+++ source/extensions/isaacsim.asset.importer.mjcf.ui/python/impl/option_widget.py
@@ -19,9 +19,9 @@ from collections.abc import Callable
 
 import omni.ui as ui
 from isaacsim.asset.importer.mjcf.impl import MJCFImporterConfig
+from isaacsim.gui.components.ui_utils import checkbox_builder, dropdown_builder, string_filed_builder
 
 from .style import get_option_style
-from .ui_utils import checkbox_builder, dropdown_builder, string_filed_builder
 
 
 def option_header(collapsed: bool, title: str) -> None: